rtd-requirements.txt
paver
FeedParser>=5.1.3
SQLAlchemy >=0.7, <0.7.99
PyYAML
beautifulsoup4>=4.1, !=4.2.0, <4.4
html5lib>=0.11
PyRSS2Gen
pynzb
progressbar
jinja2
flask
cherrypy
requests>=1.0, <2.99
python-dateutil!=2.0, !=2.2
jsonschema>=2.0
python-tvrage
tmdb3
guessit>=0.9.3
rpyc
path.py
apscheduler
